feat: create viewer interface for rendering of environments (#99)
