+++ jumanji/environments/commons/maze_utils/maze_rendering.py
@@ -24,17 +24,17 @@ from numpy.typing import NDArray
 
 import jumanji.environments
 from jumanji.environments.commons.maze_utils.maze_generation import EMPTY, WALL, Maze
+from jumanji.viewer import Viewer
 
 
-class MazeViewer:
+class MazeViewer(Viewer):
     FONT_STYLE = "monospace"
     FIGURE_SIZE = (10.0, 10.0)
     # EMPTY is white, WALL is black
     COLORS = {EMPTY: [1, 1, 1], WALL: [0, 0, 0]}
 
     def __init__(self, name: str, render_mode: str = "human") -> None:
-        """
-        Viewer for a maze environment.
+        """Viewer for a maze environment.
 
         Args:
             name: the window name to be used when initialising the window.
