+++ jumanji/environments/logic/game_2048/env.py
@@ -18,10 +18,10 @@ import chex
 import jax
 import jax.numpy as jnp
 import matplotlib.animation as animation
+from numpy.typing import NDArray
 
 from jumanji import specs
 from jumanji.env import Environment
-from jumanji.environments.logic.game_2048.env_viewer import Game2048Viewer
 from jumanji.environments.logic.game_2048.types import Board, Observation, State
 from jumanji.environments.logic.game_2048.utils import (
     move_down,
@@ -29,7 +29,9 @@ from jumanji.environments.logic.game_2048.utils import (
     move_right,
     move_up,
 )
+from jumanji.environments.logic.game_2048.viewer import Game2048Viewer
 from jumanji.types import TimeStep, restart, termination, transition
+from jumanji.viewer import Viewer
 
 
 class Game2048(Environment[State]):
@@ -78,16 +80,19 @@ class Game2048(Environment[State]):
     ```
     """
 
-    def __init__(self, board_size: int = 4) -> None:
+    def __init__(
+        self, board_size: int = 4, viewer: Optional[Viewer[State]] = None
+    ) -> None:
         """Initialize the 2048 game.
 
         Args:
             board_size: size of the board. Defaults to 4.
+            viewer: `Viewer` used for rendering. Defaults to `Game2048Viewer`.
         """
         self.board_size = board_size
 
         # Create viewer used for rendering
-        self._env_viewer = Game2048Viewer("2048", board_size)
+        self._viewer = viewer or Game2048Viewer("2048", board_size)
 
     def __repr__(self) -> str:
         """String representation of the environment.
@@ -308,15 +313,13 @@ class Game2048(Environment[State]):
         )
         return action_mask
 
-    def render(self, state: State, save_path: Optional[str] = None) -> None:
+    def render(self, state: State) -> Optional[NDArray]:
         """Renders the current state of the game board.
 
         Args:
             state: is the current game state to be rendered.
-            save_path: the path where the image should be saved. If it is None, the plot
-            will not be stored.
         """
-        return self._env_viewer.render(state=state, save_path=save_path)
+        return self._viewer.render(state=state)
 
     def animate(
         self,
@@ -335,7 +338,7 @@ class Game2048(Environment[State]):
         Returns:
             animation.FuncAnimation: the animation object that was created.
         """
-        return self._env_viewer.animate(
+        return self._viewer.animate(
             states=states, interval=interval, save_path=save_path
         )
 
@@ -345,4 +348,4 @@ class Game2048(Environment[State]):
         Environments will automatically :meth:`close()` themselves when
         garbage collected or when the program exits.
         """
-        self._env_viewer.close()
+        self._viewer.close()
