+++ jumanji/environments/routing/connector/env.py
@@ -27,7 +27,6 @@ from jumanji.environments.routing.connector.constants import (
     NOOP,
     PATH,
 )
-from jumanji.environments.routing.connector.env_viewer import ConnectorViewer
 from jumanji.environments.routing.connector.generator import (
     Generator,
     UniformRandomGenerator,
@@ -43,7 +42,9 @@ from jumanji.environments.routing.connector.utils import (
     move_position,
     switch_perspective,
 )
+from jumanji.environments.routing.connector.viewer import ConnectorViewer
 from jumanji.types import TimeStep, restart, termination, transition
+from jumanji.viewer import Viewer
 
 
 class Connector(Environment[State]):
@@ -104,7 +105,7 @@ class Connector(Environment[State]):
         generator: Optional[Generator] = None,
         reward_fn: Optional[RewardFn] = None,
         time_limit: int = 50,
-        render_mode: str = "human",
+        viewer: Optional[Viewer[State]] = None,
     ) -> None:
         """Create the `Connector` environment.
 
@@ -115,6 +116,8 @@ class Connector(Environment[State]):
             reward_fn: class of type `RewardFn`, whose `__call__` is used as a reward function.
                 Implemented options are [`DenseRewardFn`]. Defaults to `DenseRewardFn`.
             time_limit: the number of steps allowed before an episode terminates. Defaults to 50.
+            viewer: `Viewer` used for rendering. Defaults to `ConnectorViewer` with "human" render
+                mode.
         """
         self._generator = generator or UniformRandomGenerator(
             grid_size=10, num_agents=5
@@ -124,7 +127,9 @@ class Connector(Environment[State]):
         self.num_agents = self._generator.num_agents
         self.grid_size = self._generator.grid_size
         self._agent_ids = jnp.arange(self.num_agents)
-        self._renderer = ConnectorViewer("Connector", self.num_agents, render_mode)
+        self._viewer = viewer or ConnectorViewer(
+            "Connector", self.num_agents, render_mode="human"
+        )
 
     def reset(self, key: chex.PRNGKey) -> Tuple[State, TimeStep[Observation]]:
         """Resets the environment.
@@ -304,7 +309,7 @@ class Connector(Environment[State]):
         Args:
             state: `State` object containing the current environment state.
         """
-        return self._renderer.render(state.grid)
+        return self._viewer.render(state.grid)
 
     def animate(
         self,
@@ -324,7 +329,7 @@ class Connector(Environment[State]):
             animation that can export to gif, mp4, or render with HTML.
         """
         grids = [state.grid for state in states]
-        return self._renderer.animate(grids, interval, save_path)
+        return self._viewer.animate(grids, interval, save_path)
 
     def close(self) -> None:
         """Perform any necessary cleanup.
@@ -332,7 +337,7 @@ class Connector(Environment[State]):
         Environments will automatically :meth:`close()` themselves when
         garbage collected or when the program exits.
         """
-        self._renderer.close()
+        self._viewer.close()
 
     def observation_spec(self) -> specs.Spec[Observation]:
         """Specifications of the observation of the `Connector` environment.
