+++ jumanji/environments/routing/cvrp/viewer.py
@@ -12,25 +12,30 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from itertools import groupby
 from typing import Callable, Optional, Sequence, Tuple
 
 import matplotlib.animation
 import matplotlib.pyplot as plt
 import numpy as np
+from chex import Array
 from numpy.typing import NDArray
 
 import jumanji.environments
-from jumanji.environments.routing.tsp.types import State
+from jumanji.environments.routing.cvrp.types import State
+from jumanji.viewer import Viewer
 
 
-class TSPViewer:
+class CVRPViewer(Viewer):
     FIGURE_SIZE = (10.0, 10.0)
-    NODE_COLOUR = "dimgray"
+    NODE_COLOUR = "black"
+    COLORMAP_NAME = "hsv"
     NODE_SIZE = 150
+    DEPOT_SIZE = 250
     ARROW_WIDTH = 0.004
 
-    def __init__(self, name: str, render_mode: str = "human") -> None:
-        """Viewer for the TSP environment.
+    def __init__(self, name: str, num_cities: int, render_mode: str = "human") -> None:
+        """Viewer for the `CVRP` environment.
 
         Args:
             name: the window name to be used when initialising the window.
@@ -39,6 +44,10 @@ class TSPViewer:
                 - "rgb_array": return a numpy array frame representing the environment.
         """
         self._name = name
+        self._num_cities = num_cities
+
+        # Each route to and from depot has a different color
+        self._cmap = matplotlib.cm.get_cmap(self.COLORMAP_NAME, self._num_cities)
 
         # The animation must be stored in a variable that lives as long as the
         # animation should run. Otherwise, the animation will get garbage-collected.
@@ -53,7 +62,7 @@ class TSPViewer:
             raise ValueError(f"Invalid render mode: {render_mode}")
 
     def render(self, state: State) -> Optional[NDArray]:
-        """Render the given state of the `TSP` environment.
+        """Render the given state of the `CVRP` environment.
 
         Args:
             state: the environment state to render.
@@ -136,30 +145,78 @@ class TSPViewer:
         map_img = plt.imread("docs/img/city_map.jpeg")
         ax.imshow(map_img, extent=[0, 1, 0, 1])
 
+    def _group_tour(self, tour: Array) -> list:
+        """Group the tour into routes that either (1) start and end at the depot, or, (2) start at
+        the depot and end at the current city.
+
+        Args:
+            tour: x and y coordinates of the cities in the tour.
+
+        Returns:
+            tour_grouped: list of x and y coordinates that are grouped based on the above.
+        """
+        depot = tour[0]
+        check_depot_fn = lambda x: (x != depot).all()
+        tour_grouped = [
+            np.array([depot] + list(g) + [depot])
+            for k, g in groupby(tour, key=check_depot_fn)
+            if k
+        ]
+        if (tour[-1] != tour[0]).all():
+            tour_grouped[-1] = tour_grouped[-1][:-1]
+        return tour_grouped
+
+    def _draw_route(self, ax: plt.Axes, coords: Array, col_id: int) -> None:
+        """Draw the arrows and nodes for each route in the given colour."""
+        x, y = coords.T
+
+        # Compute the difference in the x- and y-coordinates to determine the distance between
+        # consecutive cities.
+        dx = x[1:] - x[:-1]
+        dy = y[1:] - y[:-1]
+        ax.quiver(
+            x[:-1],
+            y[:-1],
+            dx,
+            dy,
+            scale_units="xy",
+            angles="xy",
+            scale=1,
+            width=self.ARROW_WIDTH,
+            headwidth=5,
+            color=self._cmap(col_id),
+        )
+        ax.scatter(x, y, s=self.NODE_SIZE, color=self._cmap(col_id))
+
     def _add_tour(self, ax: plt.Axes, state: State) -> None:
-        """Add all the cities and the current tour between the visited cities to the plot."""
+        """Add the cities and the depot to the plot, and draw each route in the tour in a different
+        colour. The tour is the entire trajectory between the visited cities and a route is a
+        trajectory either starting and ending at the depot or starting at the depot and ending at
+        the current city."""
         x_coords, y_coords = state.coordinates.T
 
-        # Draw the cities as nodes
-        ax.scatter(x_coords, y_coords, s=self.NODE_SIZE, color=self.NODE_COLOUR)
+        # Draw the cities
+        ax.scatter(x_coords[1:], y_coords[1:], s=self.NODE_SIZE, color=self.NODE_COLOUR)
 
         # Draw the arrows between cities
-        if state.num_visited > 1:
-            xs, ys = state.coordinates[state.trajectory[: state.num_visited]].T
-            dx = xs[1:] - xs[:-1]
-            dy = ys[1:] - ys[:-1]
-            ax.quiver(
-                xs[:-1],
-                ys[:-1],
-                dx,
-                dy,
-                scale_units="xy",
-                angles="xy",
-                scale=1,
-                width=self.ARROW_WIDTH,
-                headwidth=5,
-            )
-            ax.scatter(xs, ys, s=self.NODE_SIZE, color="black")
+        if state.num_total_visits > 1:
+            coords = state.coordinates[state.trajectory[: state.num_total_visits]]
+            coords_grouped = self._group_tour(coords)
+
+            # Draw each route in different colour
+            for coords_route, col_id in zip(
+                coords_grouped, np.arange(0, len(coords_grouped))
+            ):
+                self._draw_route(ax, coords_route, col_id)
+
+        # Draw the depot node
+        ax.scatter(
+            x_coords[0],
+            y_coords[0],
+            marker="s",
+            s=self.DEPOT_SIZE,
+            color=self.NODE_COLOUR,
+        )
 
     def _display_human(self, fig: plt.Figure) -> None:
         if plt.isinteractive():
