+++ jumanji/environments/routing/maze/viewer.py
@@ -20,6 +20,7 @@ from numpy.typing import NDArray
 from jumanji.environments.commons.maze_utils.maze_generation import Maze
 from jumanji.environments.commons.maze_utils.maze_rendering import MazeViewer
 from jumanji.environments.routing.maze.types import State
+from jumanji.viewer import Viewer
 
 EMPTY = 0
 WALL = 1
@@ -35,7 +36,12 @@ class MazeEnvViewer(MazeViewer):
         TARGET: [1, 0, 0],  # Red
     }
 
-    def __init__(self, name: str, render_mode: str = "human") -> None:
+    def __init__(
+        self,
+        name: str,
+        render_mode: str = "human",
+        viewer: Optional[Viewer[State]] = None,
+    ) -> None:
         """Viewer for the Maze environment.
 
         Args:
